+++ learning/model/__init__.py
@@ -1,1 +1,1 @@
-from .squeezenet import Squeezenet
+from .dronet import Dronet
